requirements.txt
streamlit
pillow
numpy
plotly
opencv-python 
pyusb  
ffmpeg-python